add user login and signup and add linters

+++ server/api/user.py
@@ -6,12 +6,16 @@ import logging
 import traceback
 import mysql.connector as mysql
 
-day_trip_user = Blueprint("day_trip_user", __name__, template_folder="../client")
+day_trip_user = Blueprint("day_trip_user",
+                          __name__,
+                          template_folder="../client")
+
 
 @response.json_response
 @day_trip_user.route("/api/user", methods=["GET"])
 def get_user():
-    if 'id' in session and session.get("user_status", "not_yet_log_in") == "already_logged_in":
+    if 'id' in session and session.get(
+            "user_status", "not_yet_log_in") == "already_logged_in":
 
         user_info = user.get_user_info(session['id'])
         if user_info:
@@ -25,12 +29,16 @@ def get_user():
 def user_login():
     header_content_type = request.headers.get("Content-Type", None)
 
-    if 'id' in session and session.get("user_status", "not_yet_log_in") == "already_logged_in":
+    if 'id' in session and session.get(
+            "user_status", "not_yet_log_in") == "already_logged_in":
         return {"ok": True}, 200
 
     if header_content_type != "application/json":
         session["user_status"] = "not_yet_log_in"
-        return {"error": True, "message": "Content-type is not acceptable"}, 406
+        return {
+            "error": True,
+            "message": "Content-type is not acceptable"
+        }, 406
 
     try:
         body_info = request.get_json()
@@ -43,7 +51,8 @@ def user_login():
             if not user_id:
                 return {"error": True, "message": "User not found"}, 400
 
-            session['id'], session['user_status'] = user_id, "already_logged_in"
+            session['id'], session[
+                'user_status'] = user_id, "already_logged_in"
             return {"ok": True}, 200
 
         elif not email or not password:
@@ -61,16 +70,23 @@ def user_signup():
 
     if header_content_type != "application/json":
         session["user_status"] = "not_yet_log_in"
-        return {"error": True, "message": "Content-type is not acceptable"}, 406
+        return {
+            "error": True,
+            "message": "Content-type is not acceptable"
+        }, 406
 
     try:
         body_info = request.get_json()
-        name, email, password = body_info['name'], body_info['email'], body_info['password']
+        name, email, password = body_info['name'], body_info[
+            'email'], body_info['password']
 
         if name and email and password:
 
-            affected_rows = user.add_user(
-                {"name": name, "email": email, "password": password})
+            affected_rows = user.add_user({
+                "name": name,
+                "email": email,
+                "password": password
+            })
 
             if affected_rows > 0:
                 return {"ok": True}, 200
@@ -91,7 +107,10 @@ def user_signup():
         elif err.errno == mysql.errorcode.ER_PARSE_ERROR:
             return {"error": True, "message": "Invalid SQL syntax"}, 500
 
-        return {"error": True, "message": f"MySql error code: {mysql.errorcode}"}, 500
+        return {
+            "error": True,
+            "message": f"MySql error code: {mysql.errorcode}"
+        }, 500
     except:
         logging.error(traceback.format_exc())
         return {"error": True, "message": "Internal Server Error"}, 500
@@ -104,10 +123,15 @@ def sign_out():
     session.pop("id", None)
     return {"ok": True}, 200
 
+
 @response.json_response
 @day_trip_user.route('/api/user/delete', methods=["DELETE"])
 def user_delete_account():
-    if 'id' in session and session.get("user_status", "not_yet_log_in") == "already_logged_in":
+    if 'admin' not in session:
+        return {"error": True, "message": "You are not admin"}, 401
+
+    if 'admin' in session and 'id' in session and session.get(
+            "user_status", "not_yet_log_in") == "already_logged_in":
         try:
             affected_rows = user.delete_user(session.get("id", None))
 
@@ -123,7 +147,10 @@ def user_delete_account():
             if err.errno == mysql.errorcode.ER_PARSE_ERROR:
                 return {"error": True, "message": "Invalid SQL syntax"}, 500
 
-            return {"error": True, "message": f"MySql error code: {mysql.errorcode}"}, 500
+            return {
+                "error": True,
+                "message": f"MySql error code: {mysql.errorcode}"
+            }, 500
 
         except TypeError:
             logging.error(traceback.format_exc())
